requirements.txt
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
oauthlib==3.2.2
outcome==1.2.0
Pillow==9.5.0
pkg-resources==0.0.0
PySocks==1.7.1
requests==2.30.0
requests-oauthlib==1.3.1
selenium==4.10.0
Selenium-Screenshot==2.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.1
trio-websocket==0.10.3
tweepy==4.14.0
typing-extensions==4.7.1
urllib3==2.0.2
wsproto==1.2.0
zipp==3.15.0
